requirements.txt
streamlit
pytesseract
deep-translator
python-docx
PyPDF2
openai
anthropic
reportlab
requests
python-dotenv
textwrap3